# Decorator Performance Improvements