These changes largely bring decorators "back into the fold" performance-wise; most of the improvement is about when we update the decorator context; doing it in the start of the 'next' decorator means we avoid an additional allocation, list copy, etc.